CR.FIX.24908 Changed module values to Int type. Changes for supporting Rubika PLP.

+++ CreatorCode/CreatorClient/Entities/ProductLicense.cs
@@ -68,6 +68,7 @@ namespace Client.Creator
 
         #region Properties
         #region NonBrowsable Properties
+
         [Browsable(false)]
         public byte ProductID
         {
@@ -353,7 +354,7 @@ namespace Client.Creator
                                     if (!_commLink.IsClientType(ProductID)) //Don't need to multiply non-client module values
                                     {
                                         baseValue = (module.AppInstance > 1) ? module.Value / module.AppInstance : module.Value;
-                                        module.Value = (short)(module.Value + ((value - module.AppInstance) * baseValue));
+                                        module.Value = (module.Value + ((value - module.AppInstance) * baseValue));
                                     }
                                     //if value = mod.appinstance then no need to change
                                     module.AppInstance = value; //multiply?
@@ -890,7 +891,7 @@ namespace Client.Creator
                             uint moduleValue = (plt.plState == (byte)ProductLicenseState.Trial) ? moduleSpec.moduleTrialLicense.TVal : moduleSpec.moduleDefaultLicense.TVal;                      
                             ModuleTable mt = new ModuleTable();
                             mt.ModID = (short)moduleSpec.moduleID.TVal;
-                            mt.Value = (short)moduleValue;
+                            mt.Value = (int)moduleValue;
                             mt.AppInstance = (moduleValue > 0) ? ProductConnection : (byte)0;
                             mt.ProductLicenseID = plt.ID;
                             addModuleList.Add(mt);
@@ -931,11 +932,11 @@ namespace Client.Creator
                     int multiplier = _commLink.IsClientType(ProductID) ? 1 : ProductConnection;
                     foreach (ModuleTable module in ModuleList)
                     {
-                        short moduleValue = _commLink.GetDefaultModuleValue(ProductID, module.ModID);
+                        uint moduleValue = _commLink.GetDefaultModuleValue(ProductID, module.ModID);
                         ModuleTable mt = client.GetModule(ID, module.ModID);
                         if (mt != null)
                         {
-                            mt.Value = (short)(moduleValue * multiplier);
+                            mt.Value = (int)(moduleValue * multiplier);
                             if (moduleValue == 0) mt.AppInstance = 0;
                             mtList.Add(mt);
                         }
@@ -965,7 +966,7 @@ namespace Client.Creator
                             ModuleTable mt = client.GetModule(ID, module.ModID);
                             if (mt != null)
                             {
-                                mt.Value = (short)(_commLink.GetModuleTrialValue(ProductID, module.ModID) * multiplier);                                
+                                mt.Value = (int)(_commLink.GetModuleTrialValue(ProductID, module.ModID) * multiplier);                                
                                 mt.AppInstance = (byte)((mt.Value == 0) ? 0 : ProductConnection);
                                 mtList.Add(mt);
                             }
@@ -1212,7 +1213,7 @@ namespace Client.Creator
                                                           string.Format("Modify {0} - {1}", ProductName, module.Name),
                                                           module.Value.ToString(),
                                                           previousValue);
-                    mt.Value = (module.Value > 0) ? module.Value : (short)0;
+                    mt.Value = (module.Value > 0) ? module.Value : 0;
                     mt.AppInstance = (module.Value > 0) ? module.AppInstance : (byte)0;
                     client.UpdateModule(mt);
                     client.MarkDirty(LicenseServer);
@@ -1254,7 +1255,7 @@ namespace Client.Creator
         bool IsTrialModuleValues()
         {
             bool bRetVal = true;
-            short moduleValue;
+            uint moduleValue;
             List<ModuleTable> mtList = new List<ModuleTable>();
             foreach (ModuleTable module in ModuleList)
             {
